fix(agent): make the FreeBSD build compile, and keep it that way
GOOS=freebsd has not compiled for some time. The Backend interface gained
ListGroups(string) ([]uint32, error), and the FreeBSD backend was left holding
a stub returning ([]string, error), so the init() assigning it to
DefaultBackend did not type check:

    auth_freebsd.go:50:19: cannot use &backend{} as Backend value:
      wrong type for method ListGroups
      have ListGroups(string) ([]string, error)
      want ListGroups(string) ([]uint32, error)

Correcting the signature alone would have produced a second failure:
command_native.go is built on FreeBSD and calls osauth.ListGroups, which lived
in auth.go behind //go:build !freebsd along with the whole group file reader.

A Unix group file is the same colon-separated format everywhere the agent
runs, so none of that code was platform-specific to begin with — it was only
sitting in a file that happened to carry a tag. group.go now holds
DefaultGroupFilename, Group, parseGroupLine, parseGroupReader,
ListGroupsFromFile, parseUint32, and both ListGroups entry points, with no
build tag. The method is declared on *backend, which each platform file
defines exactly once, so it binds to whichever is compiled.

FreeBSD gets a working implementation out of it rather than the stub's
nil, nil, which would have compiled while silently giving every session no
supplementary groups.

The group tests move to group_test.go, also untagged, so GOOS=freebsd go vet
type checks them. They were in auth_test.go, which is //go:build !freebsd.

The new cross-compile job builds and vets the agent for FreeBSD on every PR
touching agent/ or pkg/. The binaries cannot run on the runner, so it is a
compile and type check — which is precisely what was missing: nothing in CI
or in the release matrix, which is Linux only, ever named FreeBSD, so an
interface change could not be noticed from the source it broke.

Unrelated and left alone: the "Code format [Go]" step in the same workflow is
guarded on matrix.os, which the validate matrix does not define, so it has
never run. gofmt -s -l is clean repo-wide today.
