+++ agent/pkg/osauth/auth_freebsd.go
@@ -42,10 +42,6 @@ func (b *backend) LookupUser(username string) (*User, error) {
 	return LookupUserFromPasswd(username, file)
 }
 
-func (b *backend) ListGroups(username string) ([]string, error) {
-	return nil, nil
-}
-
 func init() {
 	DefaultBackend = &backend{}
 }
